rename `features` folder to `entities`
