+++ src/components/categories/component.tsx
@@ -3,7 +3,7 @@ import classes from "./styles.module.scss";
 import { useState } from "react";
 import classNames from "classnames";
 import { useAppSelector } from "../redux/hooks";
-import { selectRestaurantIds } from "../redux/features/restaurant/selectors";
+import { selectRestaurantIds } from "../redux/entities/restaurant/selectors";
 
 interface Props {
   onCategorySelect: (id: string) => void;
