+++ src/components/category/component.tsx
@@ -2,7 +2,7 @@ import classes from "./styles.module.scss";
 import classNames from "classnames";
 import { Button } from "../button/component";
 import { useAppSelector } from "../redux/hooks";
-import { selectRestaurantById } from "../redux/features/restaurant/selectors";
+import { selectRestaurantById } from "../redux/entities/restaurant/selectors";
 
 interface Props {
   categoryId: string;
