+++ src/components/redux/index.ts
@@ -1,8 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { restaurantSlice } from "./features/restaurant";
-import { dishSlice } from "./features/dish";
-import { reviewSlice } from "./features/review";
-import { userSlice } from "./features/user";
+import { restaurantSlice } from "./entities/restaurant";
+import { dishSlice } from "./entities/dish";
+import { reviewSlice } from "./entities/review";
+import { userSlice } from "./entities/user";
 
 const store = configureStore({
   reducer: {
