+++ src/components/restaurant/component.tsx
@@ -4,7 +4,7 @@ import classes from "./styles.module.scss";
 import { ReviewForm } from "../review-form/component";
 import classNames from "classnames";
 import { useAppSelector } from "../redux/hooks";
-import { selectRestaurantById } from "../redux/features/restaurant/selectors";
+import { selectRestaurantById } from "../redux/entities/restaurant/selectors";
 
 interface Props {
   id: string;
